refactor: 리뷰 레포지토리 관련 테스트코드 의존성 수정
